requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
fake-useragent>=0.1.11
pyyaml>=5.4.1 